사이드바 카테고리 메뉴리스트에 적용 커밋
