+++ client/src/pages/KurlyMain.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React, { useContext, useEffect} from 'react';
 import PromoBannerSlider from '../component/main/PromoBannerSlider.jsx';
 import MainSideBarMenu from'../component/main/MainSideBarMenu.jsx';
 import MainSection from'../component/main/MainSection.jsx';
@@ -6,17 +6,7 @@ import MainPopUp from'../component/main/MainPopUp.jsx';
 
 export default function KurlyMain() {
 
-  useEffect(() => {
-   const handleScroll = () => {
-    (window.scrollY > 200)
-    ? document.body.classList.add('scrolled')
-    : document.body.classList.remove('scrolled');
-   };
 
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, []);
-  
   return (
     <div className='content-outline'>
       <PromoBannerSlider />
