+++ client/src/scss/common.scss
@@ -649,7 +649,7 @@
 ** 2025.02.15 김다희 작성
 ************************************* */
 
-.side_bar_outline{
+ .side_bar_outline{
   position: fixed;
   top: 540px;
   right: 20px;
